requirements.txt
flask==3.0.3
flask-login==0.6.3
supabase==2.4.6
pdfplumber==0.11.0
pandas==2.2.2
openai==1.30.1
reportlab==4.2.0
python-dotenv==1.0.1
gunicorn==22.0.0
Werkzeug==3.0.3
requests==2.32.3
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.130.0
